Control.Concurrent.STM.SaneTChan

Description

A sane TChan-like library. It's quite similar to TChan, but it separates the read end and the write end of the channel into distinct types. This elegantly solves the problem that was solved in a much more hackish way by newBroadcastTChan.

Documentation

data Queue a

A queue that can be written to.

data Endpoint a

An endpoint that can be used for reading items written to a queue.

newQueue :: STM (Queue a)

Creates a new Queue.

newEndpoint :: Queue a -> STM (Endpoint a)

Creates a new endpoint that reads from the specified queue. The endpoint initially starts out empty; items will become available once writeQueue is called to write items to the queue.

writeQueue :: Queue a -> a -> STM ()

Write an item to a queue. The item will then be available for reading from any endpoints previously created with newEndpoint.

readEndpoint :: Endpoint a -> STM a

Read the next value from the specified endpoint, retrying if no value is currently available.

tryReadEndpoint :: Endpoint a -> STM (Maybe a)

A version of readEndpoint which does not retry. Instead it returns Nothing if no value is available.

peekEndpoint :: Endpoint a -> STM a

Get the next value from the Endpoint without removing it, retrying if the endpoint is empty.

tryPeekEndpoint :: Endpoint a -> STM (Maybe a)

A version of peekEndpoint which does not retry. Instead it returns Nothing if no value is available.

unGetEndpoint :: Endpoint a -> a -> STM ()

Put a value back onto an endpoint, where it will be the next item read.

isEmptyEndpoint :: Endpoint a -> STM Bool

Returns True if the supplied Endpoint is empty.

cloneEndpoint :: Endpoint a -> STM (Endpoint a)

Clone an Endpoint: similar to newEndpoint, but the new endpoint starts with the same content available as the original endpoint.