Control.Concurrent.STM.SaneTChan

Description

A sane TChan-like library. It's quite similar to Control.Concurrent.STM.TChan, but it separates the read end and the write end of the channel into distinct types.

Documentation

data Queue a

A queue that can be written to.

Constructors

Queue (TVar (Link a))

data Endpoint a

An endpoint that can be used for reading items written to a queue.

Constructors

Endpoint (TVar (Link a))

data Item a

Constructors

Empty

Item a (Link a)

type Link a = TVar (Item a)

newQueue :: STM (Queue a)

Creates a new queue.

writeQueue :: Queue a -> a -> STM ()

Writes the specified item to the specified queue. The item will then be available for reading from any endpoints previously created with newEndpoint.

newEndpoint :: Queue a -> STM (Endpoint a)

Creates a new endpoint that reads from the specified queue. The endpoint initially starts out empty; items will become available once writeQueue is called to write items to the queue.

readEndpoint :: Endpoint a -> STM a

Reads the next value from the specified endpoint, retrying if no value is currently available.